# Optimize Scope.cacheReport