Optimize the cacheReport path for all Scopes.

Results for the added benchmark (master v HEAD):
```
benchmark                                  old ns/op      new ns/op     delta
BenchmarkScopeReporting/size1-4            79.6           42.0          -47.24%
BenchmarkScopeReporting/size10-4           201            62.3          -69.00%
BenchmarkScopeReporting/size100-4          1666           282           -83.07%
BenchmarkScopeReporting/size1000-4         16439          2547          -84.51%
BenchmarkScopeReporting/size10000-4        167362         26812         -83.98%
BenchmarkScopeReporting/size100000-4       4239618        328205        -92.26%
BenchmarkScopeReporting/size1000000-4      69189550       5549203       -91.98%
BenchmarkScopeReporting/size10000000-4     2205450363     35460112      -98.39%
```

Map iteration << Slice iteration